% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_concatenation.R
\name{combine_concatenation}
\alias{combine_concatenation}
\title{Combine features by concatenation}
\usage{
combine_concatenation(data, cmb_list)
}
\arguments{
\item{data}{abc}

\item{cmb_list}{abc}
}
\description{
Combine features by concatenation
}
