% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_grouping.R
\name{combine_grouping}
\alias{combine_grouping}
\title{Combine features by grouping statistics}
\usage{
combine_grouping(data, vars, group_vars)
}
\arguments{
\item{data}{A data.table with features to combine}

\item{vars}{Character vector features to combine, must be column names in data}

\item{group_vars}{Character vector with features to calculate grouping statistics, must be column names of numeric variables in data}
}
\description{
Combine features by grouping statistics
}
