% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_frequency.R
\name{encode_frequency}
\alias{encode_frequency}
\title{Frequency encoding}
\usage{
encode_frequency(data, target_var,
  vars_to_encode = colnames(data)[-match(target_var, colnames(data))],
  relative = F)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}

\item{vars_to_encode}{abc}

\item{relative}{abc}
}
\description{
Frequency encoding
}
