% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_distributions.R
\name{plot_feature_distributions}
\alias{plot_feature_distributions}
\title{Plotting function for ridgeline plot of features against target}
\usage{
plot_feature_distributions(data, target_var)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}
}
\description{
Plotting function for ridgeline plot of features against target
}
