% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_noise_feature.R
\name{plot_noise_feature}
\alias{plot_noise_feature}
\title{Plot to evaluate variable robustness of single variable}
\usage{
plot_noise_feature(data, target_var, var, folds)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}

\item{var}{abc}

\item{folds}{abc}
}
\description{
Plot to evaluate variable robustness of single variable
}
