% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rank_combinations.R
\name{plot_rank_combinations}
\alias{plot_rank_combinations}
\title{Plotting function for evaluating feature combinations by concatenation}
\usage{
plot_rank_combinations(data, target_var, top_n = 10)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}

\item{top_n}{abc}
}
\description{
Plotting function for evaluating feature combinations by concatenation
}
