% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot to evaluate feature informativeness}
\usage{
plot_signal(data, target_var, top_n = 20)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}

\item{top_n}{abc}
}
\description{
Plot to evaluate feature informativeness
}
