% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_noise.R
\name{tabulate_noise}
\alias{tabulate_noise}
\title{Evaluate variable robustness of multiple variables}
\usage{
tabulate_noise(data, target_var, vars = colnames(data)[-match(target_var,
  colnames(data))], folds)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}

\item{vars}{abc}

\item{folds}{abc}
}
\description{
Evaluate variable robustness of multiple variables
}
