% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_signal.R
\name{tabulate_signal}
\alias{tabulate_signal}
\title{Table to evaluate feature informativeness}
\usage{
tabulate_signal(data, target_var)
}
\arguments{
\item{data}{abc}

\item{target_var}{abc}
}
\description{
Table to evaluate feature informativeness
}
